feature: paciente service part 1
