+++ src/main/java/BackendC3/ClinicaOdontologica/entity/Paciente.java
@@ -10,6 +10,7 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
+@Table(name = "pacientes")
 public class Paciente {
 
     @Id
