+++ src/main/java/BackendC3/ClinicaOdontologica/entity/Turno.java
@@ -13,6 +13,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
+@Table(name = "turnos")
 public class Turno {
 
     @Id
