+++ src/main/java/BackendC3/ClinicaOdontologica/service/IPacienteService.java
@@ -1,6 +1,7 @@
 package BackendC3.ClinicaOdontologica.service;
 
+import BackendC3.ClinicaOdontologica.dto.PacienteDto;
 import BackendC3.ClinicaOdontologica.entity.Paciente;
 
-public interface IPacienteService extends ICrudService<Paciente, Integer>{
+public interface IPacienteService extends ICrudService<PacienteDto, Integer>{
 }
